Rename things and expose public API
