+++ projects/natural/src/lib/modules/file/file-select.directive.ts
@@ -1,0 +1,12 @@
+import {Directive, Input} from '@angular/core';
+import {NaturalAbstractFile} from './abstract-file';
+
+@Directive({
+    selector: '[naturalFileSelect]',
+})
+export class NaturalFileSelectDirective extends NaturalAbstractFile {
+    /**
+     * Override parent to enable it by default
+     */
+    @Input() public selectable = true;
+}
