--- projects/natural/src/lib/modules/file/ngfSelect.directive.ts
@@ -1,9 +1,0 @@
-import {Directive, Input} from '@angular/core';
-import {ngf} from './ngf.directive';
-
-@Directive({
-    selector: '[ngfSelect]',
-})
-export class ngfSelect extends ngf {
-    @Input() public selectable: any = true;
-}
