+++ projects/natural/src/lib/modules/file/public-api.ts
@@ -1,0 +1,7 @@
+/*
+ * Public API Surface of natural
+ */
+
+export * from './file-drop.directive';
+export * from './file-select.directive';
+export * from './file.module';
